Reorganize scrapers for clear delineations
